Finish Refactor
