+++ go.mod
@@ -6,4 +6,4 @@ require github.com/go-chi/chi v1.5.4
 
 require github.com/google/uuid v1.3.0
 
-require github.com/joho/godotenv v1.5.1 // indirect
+require github.com/joho/godotenv v1.5.1
